requirements.txt
flask
opencv-python
torch==2.3.1
transformers==4.37.2
Pillow

